Build out test suite
The test suite now covers all of the functionality except LDAP, which
is going to be hard to test due to the integration with external
services. Not all of the tests pass, but that reflects problems in the
existing code; the intent of the code is clear and the tests mirror
that intent. This is enough to support upgrades!