Upgrade to rails 4.0
This is the most conservative update path that gets me to 4.0 without
any new test failures. (There are still a handful of existing failures,
which I will fix when we get up to modern Rails.)

This also updates the ruby version; some dependencies were not building
and it didn't seem worth troubleshooting them. 2.1->2.3 is the smallest
update not requiring me to install another version of ruby on my system
(the 2.1 install was hairy). Earlier I tried just going straight to 2.7,
but too many things broke, so we'll have to take both the Rails and the
ruby upgrades stepwise.