Speed up URL creation
The original semantics of generate_url were broken (the intent was to
add a random suffix until it worked, but in fact it was adding a string
to an integer, thereby casting everything to int). I updated the
semantics in an earlier commit to match their intent, but this made
generate_url much slower, as it needed to make many more database calls
to ensure that it hadn't duplicated something. Not an issue in dev/test,
but a big issue in prod.

However, the original semantics -- basically trying to increment
existing URLs -- don't actually make sense. Why bother? Instead, just
randomly generate a shortcode that probably doesn't conflict with
anything in the db.

While we're at it, add rack-mini-profiler so that we can *tell* if this
helps.