Add Devise for authentication
We can't reliably log out when we've logged in with http basic, as the
browser caches the credentials. So we need a different authentication
strategy. Time to install a standard rather than using our own.

That said, LDAP is not the norm for devise, so there are still some
shaved yaks in here.